requirements.txt
flask
crewai
langchain_community
python-dotenv